# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "S54 OpenApi Documentation",
        "description": "S54 OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "S54 Lottery API Server"
        }
    ],
    "paths": {
        "/api/lottery/{number}/{country}": {
            "get": {
                "tags": [
                    "LotteryResults"
                ],
                "summary": "Get lottery results",
                "description": "Returns lottery results",
                "operationId": "getLotteryResults",
                "parameters": [
                    {
                        "name": "number",
                        "in": "path",
                        "description": "Lottery Number",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "country",
                        "in": "path",
                        "description": "Country of Origin",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LotteryResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "LotteryResponse": {
                "title": "S54 Lottery",
                "description": "Lottery Response model",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "status",
                        "type": "string",
                        "format": "string",
                        "example": "0"
                    },
                    "message": {
                        "title": "message",
                        "description": "message",
                        "type": "string",
                        "format": "string",
                        "example": "Congratulations ..."
                    },
                    "winning_numbers": {
                        "title": "winning_numbers",
                        "description": "winning numbers",
                        "type": "string",
                        "format": "array",
                        "example": "[1, 2, 3]"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "LotteryResponse"
                }
            }
        }
    },
    "tags": [
        {
            "name": "S54 Lottery API",
            "description": "API Endpoints of Projects"
        }
    ],
    "security": [
        []
    ]
}